Added ChatScreen
chat screen is navigable
