+++ lib/main.dart
@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 // firebase
 import 'package:firebase_core/firebase_core.dart'; // firebase core
 import 'package:medipal/pages/appointment.dart';
+import 'package:medipal/pages/chat.dart';
 import 'package:medipal/pages/chat_list.dart';
 import 'package:medipal/pages/dashboard.dart';
 import 'package:medipal/pages/signup.dart';
@@ -58,6 +59,7 @@ class MyApp extends StatelessWidget {
         '/AppointmentPage': (context) => AppointmentPage(),
         '/ChatListPage': (context) => ChatListPage(),
         '/Dashboard': (context) => Dashboard(),
+        '/chat': (context) => ChatScreen(receiverUserEmail: ''),
       },
     );
   }
